+++ source/extensions/filters/network/redis_proxy/command_splitter_impl.h
@@ -14,7 +14,7 @@
 #include "common/common/utility.h"
 #include "common/singleton/const_singleton.h"
 
-#include "extensions/filters/network/common/redis/client.h"
+#include "extensions/filters/network/common/redis/client_impl.h"
 #include "extensions/filters/network/redis_proxy/command_splitter.h"
 #include "extensions/filters/network/redis_proxy/conn_pool.h"
 
@@ -88,11 +88,11 @@ protected:
 /**
  * SingleServerRequest is a base class for commands that hash to a single backend.
  */
-class SingleServerRequest : public SplitRequestBase, public Common::Redis::PoolCallbacks {
+class SingleServerRequest : public SplitRequestBase, public Common::Redis::Client::PoolCallbacks {
 public:
   ~SingleServerRequest();
 
-  // Common::Redis::PoolCallbacks
+  // Common::Redis::Client::PoolCallbacks
   void onResponse(Common::Redis::RespValuePtr&& response) override;
   void onFailure() override;
 
@@ -105,7 +105,7 @@ protected:
       : SplitRequestBase(command_stats, time_source), callbacks_(callbacks) {}
 
   SplitCallbacks& callbacks_;
-  Common::Redis::PoolRequest* handle_{};
+  Common::Redis::Client::PoolRequest* handle_{};
 };
 
 /**
@@ -152,10 +152,10 @@ protected:
   FragmentedRequest(SplitCallbacks& callbacks, CommandStats& command_stats, TimeSource& time_source)
       : SplitRequestBase(command_stats, time_source), callbacks_(callbacks) {}
 
-  struct PendingRequest : public Common::Redis::PoolCallbacks {
+  struct PendingRequest : public Common::Redis::Client::PoolCallbacks {
     PendingRequest(FragmentedRequest& parent, uint32_t index) : parent_(parent), index_(index) {}
 
-    // Common::Redis::PoolCallbacks
+    // Common::Redis::Client::PoolCallbacks
     void onResponse(Common::Redis::RespValuePtr&& value) override {
       parent_.onChildResponse(std::move(value), index_);
     }
@@ -163,7 +163,7 @@ protected:
 
     FragmentedRequest& parent_;
     const uint32_t index_;
-    Common::Redis::PoolRequest* handle_{};
+    Common::Redis::Client::PoolRequest* handle_{};
   };
 
   virtual void onChildResponse(Common::Redis::RespValuePtr&& value, uint32_t index) PURE;
