+++ source/extensions/filters/network/redis_proxy/config.cc
@@ -8,6 +8,7 @@
 
 #include "common/config/filter_json.h"
 
+#include "extensions/filters/network/common/redis/client_impl.h"
 #include "extensions/filters/network/common/redis/codec_impl.h"
 #include "extensions/filters/network/redis_proxy/command_splitter_impl.h"
 #include "extensions/filters/network/redis_proxy/conn_pool_impl.h"
@@ -30,7 +31,7 @@ Network::FilterFactoryCb RedisProxyFilterConfigFactory::createFilterFactoryFromP
       proto_config, context.scope(), context.drainDecision(), context.runtime()));
   ConnPool::InstancePtr conn_pool(new ConnPool::InstanceImpl(
       filter_config->cluster_name_, context.clusterManager(),
-      ConnPool::ClientFactoryImpl::instance_, context.threadLocal(), proto_config.settings()));
+      Common::Redis::Client::ClientFactoryImpl::instance_, context.threadLocal(), proto_config.settings()));
   std::shared_ptr<CommandSplitter::Instance> splitter(new CommandSplitter::InstanceImpl(
       std::move(conn_pool), context.scope(), filter_config->stat_prefix_, context.timeSource()));
   return [splitter, filter_config](Network::FilterManager& filter_manager) -> void {
