+++ source/extensions/filters/network/redis_proxy/conn_pool.h
@@ -31,9 +31,9 @@ public:
    * @return PoolRequest* a handle to the active request or nullptr if the request could not be made
    *         for some reason.
    */
-  virtual Common::Redis::PoolRequest* makeRequest(const std::string& hash_key,
+  virtual Common::Redis::Client::PoolRequest* makeRequest(const std::string& hash_key,
                                                   const Common::Redis::RespValue& request,
-                                                  Common::Redis::PoolCallbacks& callbacks) PURE;
+                                                  Common::Redis::Client::PoolCallbacks& callbacks) PURE;
 };
 
 typedef std::unique_ptr<Instance> InstancePtr;
